+++ src/main/java/nus/iss/se/magicbag/repository/MagicBagRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import nus.iss.se.magicbag.entity.CartItem;
 import nus.iss.se.magicbag.entity.MagicBag;
 
-public interface MagicBagRepository extends JpaRepository<MagicBag, Long> {
+public interface MagicBagRepository extends JpaRepository<MagicBag, Integer> {
 
 }
